metadata-services/pyproject.toml
[tool.poetry]
name = "metadata-service"
version = "0.1.0"
description = "Metadata service for datasets with search and lineage"
authors = ["Vigneshchandra <vignesh@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.10"

fastapi = "^0.110.0"
uvicorn = "^0.27.0"

sqlalchemy = "^2.0.0"
pymysql = "^1.1.0"
alembic = "^1.13.0"

python-dotenv = "^1.0.0"
pydantic = "^2.6.0"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
